refactor: 스토리북 디렉토리 정리
