+++ frontend/src/stories/@common/Button.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import Button from '../components/@common/buttons/button/Button';
+import Button from '../../components/@common/buttons/button/Button';
 
 const meta: Meta<typeof Button> = {
   title: 'Components/Button/Button',
